user-service/pyproject.toml
[tool.poetry]
name = "user-service"
version = "0.1.0"
description = ""
authors = ["Sami Tahri <sismixx@hotmail.fr>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.81.0"
SQLAlchemy = "^1.4.40"
SQLAlchemy-Utils = "^0.38.3"
pytest = "^7.1.3"
python-outbox = "0.1.8"
alembic = "^1.8.1"
psycopg2-binary = "2.8.6"
passlib = "^1.7.4"
uvicorn = {extras = ["standard"], version = "^0.18.3"}
PyJWT = "^2.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
